requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
streamlit
groq
langchain-groq
langchain-core
python-dotenv
joblib
